multimodel/requirements.txt
flask==2.0.1
faiss-cpu==1.7.2
google-generativeai==0.3.0
python-dotenv==0.19.0
PyPDF2==3.0.1
numpy==1.21.2
streamlit
PyPDF2
scikit-learn
pytesseract
Pillow
google-generativeai